==> 2 Clustering local.ipynb <==
# Import modules

import sys
import pandas as pd
import numpy as np
import random
import pickle
import datetime
import matplotlib.pyplot as plt
%matplotlib inline

%run MyFunctions.ipynb

# Read donation data (relative time)
df = pd.read_pickle('X:/df_rel_full.pkl')
df = df.loc[df['Donatiesoortcode'] == 'V', ]
df['TimeSinceFirst'] = df['TimeSinceFirst']

# Read donor data
df_donors = pd.read_pickle('X:/donors.pkl')

# Only take donors that have donated in 2010 or later, but not before (proxy for new donors)
old_donors = np.unique(list(df.loc[df['DateTime'].dt.year < 2010, 'KeyID']))
df = df.loc[~df['KeyID'].isin(old_donors), ]
df_donors = df_donors.loc[df_donors.index.isin(np.unique(df['KeyID'])), ]

# List of KeyIDs by sex
id_f = list(df_donors.loc[df_donors['Geslacht'] == 'V', ].index)
id_m = list(df_donors.loc[df_donors['Geslacht'] == 'M', ].index)

print("Total number of female donors:", len(id_f))
print("Total number of male donors", len(id_m))

df_f = df.loc[df['KeyID'].isin(id_f), ]
df_m = df.loc[df['KeyID'].isin(id_m), ]
countsf = df_f['KeyID'].value_counts()
countsm = df_m['KeyID'].value_counts()
id_f9 = list(np.unique(df_f.loc[df_f['KeyID'].isin(countsf.index[countsf > 8]), 'KeyID'])) 
id_m9 = list(np.unique(df_m.loc[df_m['KeyID'].isin(countsm.index[countsm > 8]), 'KeyID'])) 

print("Number of female donors with 9 or more donations in 9 years:", len(id_f9))
print("Number of male donors with 9 or more donations in 9 years:", len(id_m9))

# df_weekly_f = makeSeriesDf(data=df, ids=id_f9, interval=7, xmax=9*365)
# df_weekly_f.to_pickle('X:/df_weekly_f.pkl')
# df_weekly_m = makeSeriesDf(data=df, ids=id_m9, interval=7, xmax=9*365)
# df_weekly_m.to_pickle('X:/df_weekly_m.pkl')

df_weekly_f = pd.read_json('X:/df_weekly_f.json')
df_weekly_m = pd.read_json('X:/df_weekly_m.json')

df_weekly_f.drop(df_weekly_f.columns[len(df_weekly_f.columns)-50:], axis=1, inplace=True)
df_weekly_m.drop(df_weekly_m.columns[len(df_weekly_m.columns)-50:], axis=1, inplace=True)

df_weekly_f.to_json('X:/df_weekly_f.json')
df_weekly_m.to_json('X:/df_weekly_m.json')

# n = 5000
# selection_f = random.sample(list(df_weekly_f.index), n)
# selection_m = random.sample(list(df_weekly_m.index), n)

# df_weekly_f.loc[selection_f].to_pickle('X:/df_weekly_f_5000.pkl')
# df_weekly_m.loc[selection_m].to_pickle('X:/df_weekly_m_5000.pkl')
# df_weekly_f.loc[selection_f].to_json('X:/df_weekly_f_5000.json')
# df_weekly_m.loc[selection_m].to_json('X:/df_weekly_m_5000.json')

# FEMALE

df_weekly_f = pd.read_json('X:/df_weekly_f_5000.json')
df_weekly_m = pd.read_json('X:/df_weekly_m_5000.json')

df_weekly_f.drop(df_weekly_f.columns[len(df_weekly_f.columns)-50:], axis=1, inplace=True)
df_weekly_m.drop(df_weekly_m.columns[len(df_weekly_m.columns)-50:], axis=1, inplace=True)

# print("Women")
# for k in [5]:
#     print('\nStarting clustering with', k, 'clusters')
#     print(datetime.datetime.now())
#     c, a, d = k_means_clust(data=df_weekly_f, num_clust=k, num_iter=5, plot=True)
#     c.to_pickle('X:/clusterresults_20190601/clusters_f'+str(k)+'.pkl')
#     d.to_pickle('X:/clusterresults_20190601/distances_f'+str(k)+'.pkl')
#     out = open('X:/clusterresults_20190601/assignments_f'+str(k)+'.pkl', 'wb')
#     pickle.dump(a, out)
#     out.close()
    
# # MALE

# print("Men")
# for k in [5]:
#     print('\nStarting clustering with', k, 'clusters')
#     print(datetime.datetime.now())
#     c, a, d = k_means_clust(data=df_weekly_m, num_clust=k, num_iter=5, plot=True)
#     c.to_pickle('X:/clusterresults_20190601/clusters_m'+str(k)+'.pkl')
#     d.to_pickle('X:/clusterresults_20190601/distances_m'+str(k)+'.pkl')
#     out = open('X:/clusterresults_20190601/assignments_m'+str(k)+'.pkl', 'wb')
#     pickle.dump(a, out)
#     out.close()

df['TimeSinceFirst'] = df['TimeSinceFirst'] * 365
df.head()

df_weekly_f.shape[0]

all_index = list(df_weekly_f.index) + list(df_weekly_m.index)

df2 = df.loc[df['KeyID'].isin(all_index), ]

x = [1, 2, 3, 4, 5, 6]
a = x[1:]
b = x[:-1]

[i - j for i, j in zip(a, b)]

print(a)
print(b)

df_f = df.loc[df['KeyID'].isin(list(df_weekly_f.index)), ]

a = df_f.loc[df_f['TimeSinceFirst'] > 0, 'TimeSinceFirst'][1:] 
b = df_f.loc[df_f['TimeSinceFirst'] > 0, 'TimeSinceFirst'][:-1]
diffs_f = [i - j for i, j in zip(a, b)]

df_m = df.loc[df['KeyID'].isin(list(df_weekly_m.index)), ]

a = df_m.loc[df_m['TimeSinceFirst'] > 0, 'TimeSinceFirst'][1:] 
b = df_m.loc[df_m['TimeSinceFirst'] > 0, 'TimeSinceFirst'][:-1]
diffs_m = [i - j for i, j in zip(a, b)]

print("Women", np.median(diffs_f), np.percentile(diffs_f, 25, interpolation='midpoint'), np.percentile(diffs_f, 75, interpolation='midpoint'))
print("Men", np.median(diffs_m), np.percentile(diffs_m, 25, interpolation='midpoint'), np.percentile(diffs_m, 75, interpolation='midpoint'))

len_f = list(df.loc[df['KeyID'].isin(list(df_weekly_f.index)), ].groupby('KeyID').count()['DateTime'])
len_m = list(df.loc[df['KeyID'].isin(list(df_weekly_m.index)), ].groupby('KeyID').count()['DateTime'])

plt.hist(len_m)

np.percentile(len_f, 75, interpolation='midpoint')

np.percentile(len_m, 75, interpolation='midpoint')
