==> 3 Clustering results.ipynb <==
# Import modules

import sys
import pandas as pd
import numpy as np
import random
import pickle
import datetime
import matplotlib.pyplot as plt
%matplotlib inline
from sklearn import preprocessing
from sklearn.cluster import KMeans

%run MyFunctions.ipynb

# Import all distances and make one big dataframe

distFL2 = pd.read_pickle('X:/201902 Hb Trajectories/clusterresultslocal/distances_f2.pkl').reset_index().rename(columns={'index':'KeyID'})
distFL3 = pd.read_pickle('X:/201902 Hb Trajectories/clusterresultslocal/distances_f3.pkl').reset_index().rename(columns={'index':'KeyID'})
distFL4 = pd.read_pickle('X:/201902 Hb Trajectories/clusterresultslocal/distances_f4.pkl').reset_index().rename(columns={'index':'KeyID'})
distFL5 = pd.read_pickle('X:/201902 Hb Trajectories/clusterresultslocal/distances_f5.pkl').reset_index().rename(columns={'index':'KeyID'})

distFG2 = pd.read_json('X:/201902 Hb Trajectories/clusterresultsglobal/distances_f2.json').reset_index().rename(columns={'index':'KeyID'})
distFG3 = pd.read_json('X:/201902 Hb Trajectories/clusterresultsglobal/distances_f3.json').reset_index().rename(columns={'index':'KeyID'})
distFG4 = pd.read_json('X:/201902 Hb Trajectories/clusterresultsglobal/distances_f4.json').reset_index().rename(columns={'index':'KeyID'})
distFG5 = pd.read_json('X:/201902 Hb Trajectories/clusterresultsglobal/distances_f5.json').reset_index().rename(columns={'index':'KeyID'})

distML2 = pd.read_pickle('X:/201902 Hb Trajectories/clusterresultslocal/distances_m2.pkl').reset_index().rename(columns={'index':'KeyID'})
distML3 = pd.read_pickle('X:/201902 Hb Trajectories/clusterresultslocal/distances_m3.pkl').reset_index().rename(columns={'index':'KeyID'})
distML4 = pd.read_pickle('X:/201902 Hb Trajectories/clusterresultslocal/distances_m4.pkl').reset_index().rename(columns={'index':'KeyID'})
distML5 = pd.read_pickle('X:/201902 Hb Trajectories/clusterresultslocal/distances_m5.pkl').reset_index().rename(columns={'index':'KeyID'})

distMG2 = pd.read_json('X:/201902 Hb Trajectories/clusterresultsglobal/distances_m2.json').reset_index().rename(columns={'index':'KeyID'})
distMG3 = pd.read_json('X:/201902 Hb Trajectories/clusterresultsglobal/distances_m3.json').reset_index().rename(columns={'index':'KeyID'})
distMG4 = pd.read_json('X:/201902 Hb Trajectories/clusterresultsglobal/distances_m4.json').reset_index().rename(columns={'index':'KeyID'})
distMG5 = pd.read_json('X:/201902 Hb Trajectories/clusterresultsglobal/distances_m5.json').reset_index().rename(columns={'index':'KeyID'})

distFL2['cluster'] = translateClusters(distFL2['cluster'], ['A', 'B'])
distFL3['cluster'] = translateClusters(distFL3['cluster'], ['C', 'A', 'B'])
distFL4['cluster'] = translateClusters(distFL4['cluster'], ['B', 'D', 'A', 'C'])
distFL5['cluster'] = translateClusters(distFL5['cluster'], ['A', 'B', 'E', 'C', 'D'])

distML2['cluster'] = translateClusters(distML2['cluster'], ['A', 'B'])
distML3['cluster'] = translateClusters(distML3['cluster'], ['B', 'C', 'A'])
distML4['cluster'] = translateClusters(distML4['cluster'], ['C', 'D', 'A', 'B'])
distML5['cluster'] = translateClusters(distML5['cluster'], ['D', 'B', 'A', 'E', 'C'])

distFG2['cluster'] = translateClusters(distFG2['cluster'], ['B', 'A'])
distFG3['cluster'] = translateClusters(distFG3['cluster'], ['B', 'C', 'A'])
distFG4['cluster'] = translateClusters(distFG4['cluster'], ['A', 'C', 'D', 'B'])
distFG5['cluster'] = translateClusters(distFG5['cluster'], ['B', 'D', 'C', 'E', 'A'])

distMG2['cluster'] = translateClusters(distMG2['cluster'], ['A', 'B'])
distMG3['cluster'] = translateClusters(distMG3['cluster'], ['B', 'C', 'A'])
distMG4['cluster'] = translateClusters(distMG4['cluster'], ['D', 'B', 'C', 'A'])
distMG5['cluster'] = translateClusters(distMG5['cluster'], ['E', 'A', 'C', 'B', 'D'])

distFL2['K'], distFL3['K'], distFL4['K'], distFL5['K'] = [2, 3, 4, 5]
distFL2['sex'], distFL3['sex'], distFL4['sex'], distFL5['sex'] = ['F', 'F', 'F', 'F']
distFL2['method'], distFL3['method'], distFL4['method'], distFL5['method'] = ['local', 'local', 'local', 'local']

distFG2['K'], distFG3['K'], distFG4['K'], distFG5['K'] = [2, 3, 4, 5]
distFG2['sex'], distFG3['sex'], distFG4['sex'], distFG5['sex'] = ['F', 'F', 'F', 'F']
distFG2['method'], distFG3['method'], distFG4['method'], distFG5['method'] = ['global', 'global', 'global', 'global']

distML2['K'], distML3['K'], distML4['K'], distML5['K'] = [2, 3, 4, 5]
distML2['sex'], distML3['sex'], distML4['sex'], distML5['sex'] = ['M', 'M', 'M', 'M']
distML2['method'], distML3['method'], distML4['method'], distML5['method'] = ['local', 'local', 'local', 'local']

distMG2['K'], distMG3['K'], distMG4['K'], distMG5['K'] = [2, 3, 4, 5]
distMG2['sex'], distMG3['sex'], distMG4['sex'], distMG5['sex'] = ['M', 'M', 'M', 'M']
distMG2['method'], distMG3['method'], distMG4['method'], distMG5['method'] = ['global', 'global', 'global', 'global']

distances = distFL2.append([distFL3, distFL4, distFL5, 
                            distFG2, distFG3, distFG4, distFG5, 
                            distML2, distML3, distML4, distML5, distMG2, distMG3, distMG4, distMG5]).reset_index(drop=True)
distances.loc[:, 'distance'] = pd.to_numeric(distances.loc[:, 'distance'])
distances.head()

# Local clustering

# Load centroids and reorder for plotting LOCAL CLUSTERS

cenf2 = pd.read_pickle('X:/201902 Hb Trajectories/clusterresultslocal/clusters_f2.pkl')
cenf3 = pd.read_pickle('X:/201902 Hb Trajectories/clusterresultslocal/clusters_f3.pkl')
cenf4 = pd.read_pickle('X:/201902 Hb Trajectories/clusterresultslocal/clusters_f4.pkl')
cenf5 = pd.read_pickle('X:/201902 Hb Trajectories/clusterresultslocal/clusters_f5.pkl')
cenf2 = cenf2
cenf3 = cenf3.loc[[2, 0, 1]].reset_index(drop=True)
cenf4 = cenf4.loc[[1, 3, 0, 2]].reset_index(drop=True)
cenf5 = cenf5.loc[[0, 1, 4, 2, 3]].reset_index(drop=True)

cenm2 = pd.read_pickle('X:/201902 Hb Trajectories/clusterresultslocal/clusters_m2.pkl')
cenm3 = pd.read_pickle('X:/201902 Hb Trajectories/clusterresultslocal/clusters_m3.pkl')
cenm4 = pd.read_pickle('X:/201902 Hb Trajectories/clusterresultslocal/clusters_m4.pkl')
cenm5 = pd.read_pickle('X:/201902 Hb Trajectories/clusterresultslocal/clusters_m5.pkl')
cenm2 = cenm2
cenm3 = cenm3.loc[[1, 2, 0]].reset_index(drop=True)
cenm4 = cenm4.loc[[2, 3, 0, 1]].reset_index(drop=True)
cenm5 = cenm5.loc[[3, 1, 0, 4, 2]].reset_index(drop=True)

plt.rcParams.update({"pgf.rcfonts": False})

colors = ['red', 'green', 'orange', 'blue', 'purple']
labels = ['Cluster A', 'Cluster B', 'Cluster C', 'Cluster D', 'Cluster E']

fig, ax = plt.subplots(2, 4, figsize=(15, 8), sharex=True, sharey=True)

for a in fig.axes:
    a.tick_params(axis='both', labelsize=8, pad=1)   
    a.grid(which='major')


# Women on the left
for index, row in cenf2.iterrows():
    ax[0, 0].plot(cenf2.loc[index][:286], color=colors[index], label=labels[index])
ax[0, 0].set_ylim([7, 11])
    
for index, row in cenf3.iterrows():
    ax[0, 1].plot(cenf3.loc[index][:286], color=colors[index], label=labels[index]) 
    
for index, row in cenf4.iterrows():
    ax[0, 2].plot(cenf4.loc[index][:286], color=colors[index], label=labels[index])
    
for index, row in cenf5.iterrows():
    ax[0, 3].plot(cenf5.loc[index][:286], color=colors[index], label=labels[index])

# Men on the right
for index, row in cenm2.iterrows():
    ax[1, 0].plot(cenm2.loc[index][:286], color=colors[index], label=labels[index])
    
for index, row in cenm3.iterrows():
    ax[1, 1].plot(cenm3.loc[index][:286], color=colors[index], label=labels[index]) 
    
for index, row in cenm4.iterrows():
    ax[1, 2].plot(cenm4.loc[index][:286], color=colors[index], label=labels[index])
    
for index, row in cenm5.iterrows():
    ax[1, 3].plot(cenm5.loc[index][:286], color=colors[index], label=labels[index])

ax[1, 3].legend(loc='lower right', bbox_to_anchor=(-3.65, -1.4, 4.7, 1), ncol=3, mode="expand", fontsize=9)

cols = ['Women', 'Men']
rows = ['k = {}'.format(k) for k in range(2, 6)]

for axis, row in zip(ax[0], rows):
    axis.set_title(row, fontsize=9)

for axis, col in zip(ax[:,0], cols):
    axis.annotate(col, xy=(0, 0.5), xytext=(-axis.yaxis.labelpad - 5, 0),
                  xycoords=axis.yaxis.label, textcoords='offset points',
                  ha='left', rotation=90, va='center')
    axis.set_ylabel('Hb (mmol/L)', fontsize=9)

plt.subplots_adjust(left=None, bottom=None, right=None, top=None, wspace=0.2, hspace=None)
fig.text(0.5, -0.05, 'Days since first donation', ha='center', fontsize=9)    

# plt.savefig('fig_clusters_local.pdf', bbox_inches="tight")
# plt.savefig('fig_clusters_local.pgf', bbox_inches="tight")

distances_local = distances.loc[distances['method'] == 'local', ]
distances_local.groupby(['sex', 'K'])['distance'].sum()

# Global clustering

with open('X:/201902 Hb Trajectories/clusterresultsglobal/centroidsf.pkl', 'rb') as f:
     centroidsf = pickle.load(f)

with open('X:/201902 Hb Trajectories/clusterresultsglobal/centroidsm.pkl', 'rb') as f:
     centroidsm = pickle.load(f)

centroidsf[0] = centroidsf[0].loc[[1, 0]].reset_index(drop=True)
centroidsf[1] = centroidsf[1].loc[[1, 2, 0]].reset_index(drop=True)
centroidsf[2] = centroidsf[2].loc[[0, 2, 3, 1]].reset_index(drop=True)
centroidsf[3] = centroidsf[3].loc[[1, 3, 2, 4, 0]].reset_index(drop=True)

centroidsm[0] = centroidsm[0]
centroidsm[1] = centroidsm[1].loc[[1, 2, 0]].reset_index(drop=True)
centroidsm[2] = centroidsm[2].loc[[3, 1, 2, 0]].reset_index(drop=True)
centroidsm[3] = centroidsm[3].loc[[4, 0, 2, 1, 3]].reset_index(drop=True)

plt.rcParams.update({"pgf.rcfonts": False})

colors = ['red', 'green', 'orange', 'blue', 'purple']
labels = ['Cluster A', 'Cluster B', 'Cluster C', 'Cluster D', 'Cluster E']

fig, ax = plt.subplots(2, 4, figsize=(15, 8), sharex=True, sharey=True)

for a in fig.axes:
    a.tick_params(axis='both', labelsize=8, pad=1)
    a.grid(which='major')

for index, row in centroidsf[0].iterrows():
    ax[0, 0].plot(centroidsf[0].loc[index][:286], color=colors[index], label=labels[index])
ax[0, 0].set_ylim([7, 11])
    
for index, row in centroidsf[1].iterrows():
    ax[0, 1].plot(centroidsf[1].loc[index][:286], color=colors[index], label=labels[index])
    
for index, row in centroidsf[2].iterrows():
    ax[0, 2].plot(centroidsf[2].loc[index][:286], color=colors[index], label=labels[index])
    
for index, row in centroidsf[3].iterrows():
    ax[0, 3].plot(centroidsf[3].loc[index][:286], color=colors[index], label=labels[index])
    
for index, row in centroidsm[0].iterrows():
    ax[1, 0].plot(centroidsm[0].loc[index][:286], color=colors[index], label=labels[index])
    
for index, row in centroidsm[1].iterrows():
    ax[1, 1].plot(centroidsm[1].loc[index][:286], color=colors[index], label=labels[index])
    
for index, row in centroidsm[2].iterrows():
    ax[1, 2].plot(centroidsm[2].loc[index][:286], color=colors[index], label=labels[index])
    
for index, row in centroidsm[3].iterrows():
    ax[1, 3].plot(centroidsm[3].loc[index][:286], color=colors[index], label=labels[index])

ax[1, 3].legend(loc='lower right', bbox_to_anchor=(-3.65, -1.4, 4.7, 1), ncol=3, mode="expand", fontsize=9)
    
cols = ['Women', 'Men']
rows = ['k = {}'.format(k) for k in range(2, 6)]

for axis, row in zip(ax[0], rows):
    axis.set_title(row, fontsize=9)

for axis, col in zip(ax[:,0], cols):
    axis.annotate(col, xy=(0, 0.5), xytext=(-axis.yaxis.labelpad - 5, 0),
                  xycoords=axis.yaxis.label, textcoords='offset points',
                  ha='left', rotation=90, va='center')
    axis.set_ylabel('Hb (mmol/L)', fontsize=9)

plt.subplots_adjust(left=None, bottom=None, right=None, top=None, wspace=0.2, hspace=None)
fig.text(0.5, 0, 'Days since first donation', ha='center', fontsize=9)    

# plt.savefig('fig_clusters_global.pdf', bbox_inches="tight")
# plt.savefig('fig_clusters_global.pgf', bbox_inches="tight")

# Within-cluster distances GLOBAL CLUSTERS

distances_global = distances.loc[distances['method'] == 'global', ]
distances_global.groupby(['sex', 'K'])['distance'].sum()

# Local&global

distances.head()

sex = 'M'
K = 4

x = distances.loc[(distances['sex'] == sex) & (distances['K'] == K), ]
x2 = x.pivot(index='KeyID', columns='method', values='cluster')
pd.crosstab(x2['local'], x2['global']).apply(lambda r: round(r/r.sum()*100), axis=1)

# Visualizing some clusters
k = 4, male

df_weekly_m = pd.read_json('X:/201902 Hb Trajectories/df_weekly_m_5000.json')
df_weekly_m.head()

distancesM4 = distances.loc[(distances['K'] == 4) &
                            (distances['sex'] == 'M')]

distancesM4.head()

n = 100
d1 = distancesM4.loc[(distancesM4['method'] == 'local') & (distancesM4['cluster'] == 'A'), ].head(n)
d2 = distancesM4.loc[(distancesM4['method'] == 'local') & (distancesM4['cluster'] == 'B'), ].head(n)
d3 = distancesM4.loc[(distancesM4['method'] == 'local') & (distancesM4['cluster'] == 'C'), ].head(n)
d4 = distancesM4.loc[(distancesM4['method'] == 'local') & (distancesM4['cluster'] == 'D'), ].head(n)
d5 = distancesM4.loc[(distancesM4['method'] == 'global') & (distancesM4['cluster'] == 'A'), ].head(n)
d6 = distancesM4.loc[(distancesM4['method'] == 'global') & (distancesM4['cluster'] == 'B'), ].head(n)
d7 = distancesM4.loc[(distancesM4['method'] == 'global') & (distancesM4['cluster'] == 'C'), ].head(n)
d8 = distancesM4.loc[(distancesM4['method'] == 'global') & (distancesM4['cluster'] == 'D'), ].head(n)

distancesM4small = d1.append([d2, d3, d4, d5, d6, d7, d8])

centroidsglobal = centroidsm[2]
centroidslocal = cenm4
# centroidslocal = centroidslocal.loc[[1, 3, 0, 2]].reset_index(drop=True)
centroidsglobal = centroidsglobal.loc[[2, 3, 0, 1]].reset_index(drop=True)

plt.rcParams.update({"pgf.rcfonts": False})

fig, ax = plt.subplots(2, 4, figsize=(15, 8), sharex=True, sharey=True)

for a in fig.axes:
    a.tick_params(axis='both', labelsize=8, pad=1)   

for index, row in distancesM4small.iterrows():
    y = 0 if row['method'] == 'local' else 1
    if y == 0:
        x = 0 if row['cluster'] == 'A' else (1 if row['cluster'] == 'B' else (2 if row['cluster'] == 'C' else 3))
    else:
        x = 0 if row['cluster'] == 'C' else (1 if row['cluster'] == 'D' else (2 if row['cluster'] == 'B' else 3))
    
    ax[y, x].plot(df_weekly_m.loc[row['KeyID']][:286], color='black', alpha=0.1, label='_nolegend_')
# ax[2, 1].plot(df_weekly_m.loc['DK.00305854'][:286], color='black', alpha=0.1, label='Individual')
    
for index, row in centroidslocal.iterrows():
    x = index
    y = 0
    ax[y, x].plot(row[:286], lw=2, color='red', label='_nolegend_')
    
for index, row in centroidsglobal.iterrows():
    x = index
    y = 1
    ax[y, x].plot(row[:286], lw=2, color='red', label='Cluster centroid')

# ax[2, 1].legend(loc='upper left', bbox_to_anchor=(-1, -1.8), ncol=2, fontsize=9)

cols = ['Direct clustering', 'Trend clustering']
rows = ['Cluster {}'.format(k) for k in ['A', 'B', 'C', 'D']]

# for axis, col in zip(ax[0], cols):
#     axis.set_title(col, fontsize=9)

# for axis, row in zip(ax[:,0], rows):
#     axis.annotate(row, xy=(0, 0.5), xytext=(-axis.yaxis.labelpad - 5, 0),
#                   xycoords=axis.yaxis.label, textcoords='offset points',
#                   ha='left', rotation=90, va='center', fontsize=9)
#     axis.set_ylabel('Hb', fontsize=9)

plt.subplots_adjust(left=None, bottom=None, right=None, top=None, wspace=0.2, hspace=None)
fig.text(0.5, 0.02, 'Days since first donation', ha='center', fontsize=9)    

# plt.savefig('fig_clusters_compare.pdf', bbox_inches="tight")
# plt.savefig('fig_clusters_compare.pgf', bbox_inches="tight")
